+++ app/src/main/res/layout/mix_recipe_multiple.xml
@@ -12,7 +12,7 @@
         <LinearLayout
                 android:layout_width="match_parent"
                 android:layout_height="wrap_content"
-                android:layout_marginTop="20dp"
+                android:layout_marginTop="40dp"
                 android:layout_marginStart="16dp"
                 android:layout_marginRight="16dp"
         >
