+++ app/src/main/res/layout/mix_recipe_single.xml
@@ -8,13 +8,17 @@
     <LinearLayout
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
-            android:orientation="horizontal">
+            android:orientation="horizontal"
+            android:layout_marginStart="16dp"
+            android:layout_marginRight="16dp"
+    >
+
         <LinearLayout
                 android:layout_width="match_parent"
                 android:layout_height="wrap_content"
-                android:layout_marginTop="20dp"
-                android:layout_marginStart="16dp"
+                android:layout_marginTop="40dp"
                 android:layout_marginRight="16dp"
+
         >
             <Button
                     android:id="@+id/BtnToHome"
