+++ app/src/main/res/layout/recipe_single_item.xml
@@ -3,11 +3,12 @@
 <LinearLayout
         xmlns:android="http://schemas.android.com/apk/res/android"
         android:layout_width="match_parent"
-        android:layout_height="match_parent"
+        android:layout_height="511dp"
         android:orientation="vertical"
         android:background="#765959"
+        android:layout_marginBottom="31dp"
 >
-    <Button
+    <ImageButton
             android:id="@+id/BtnScrap"
             android:layout_width="24dp"
             android:layout_height="24dp"
@@ -23,8 +24,7 @@
             android:layout_marginStart="20dp"
             android:textSize="20sp"
             android:textColor="#ffffff"
-            android:layout_marginTop="380dp"
-            android:layout_marginBottom="8dp"
+            android:layout_marginTop="340dp"
             android:text="Alchoholic"
     />
     <TextView
@@ -35,7 +35,6 @@
             android:textSize="30sp"
             android:textColor="#ffffff"
             android:layout_marginTop="8dp"
-            android:layout_marginBottom="32dp"
             android:text="칵테일 이름"
     />
